Fixed the ghost consumptions issue
Consumption list was being shared between Statistics & Adapter

Might call for a customer object to wrap around the List<Consumption> so that we can control access. There are 'business' rules around adding
consumptions to the list. Sometimes they are added before they have an Id, then when they come back from the database they are getting added again. This
is where we need to use the group property, which uses a Guid which is generated before persisting.